pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tinyagent"
version = "0.1.0"
description = "Local-first AI agent with smart delegation to external providers"
requires-python = ">=3.11"
dependencies = [
    "ollama",
    "anthropic",
    "openai",
    "google-genai",
    "rich",
    "prompt_toolkit",
    "tomli; python_version < '3.11'",
]

[project.scripts]
tinyagent = "tinyagent.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/tinyagent"]
